devop: refactor settings files
