+++ v2/src/modules/settings/ModuleAddressNames.vue
